add first unitest by NUnit.
