+++ UnitTests/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NUnit" version="3.10.1" targetFramework="net46" />
+  <package id="NUnit.ConsoleRunner" version="3.8.0" targetFramework="net46" />
+  <package id="NUnit3TestAdapter" version="3.10.0" targetFramework="net46" />
+</packages>
